refactor: form으로 상태 이동
